Updated the `build_pqc_server.sh` script to fix an 'Argument list too long' error when building the `liboqs` static library on Windows/MSYS2.
This commit adds several `CMAKE_*_ARCHIVE_*` flags to the `liboqs` CMake command. These flags instruct CMake to use a multi-step process for creating the static archive (create, then append files) instead of a single command with all files. This is the standard workaround for this fundamental Windows command-line length limitation.

The `BUILD_SHARED_LIBS` flag was also reverted to `OFF` to build the static library as originally intended.