Updated the `build_pqc_server.sh` script to add the `-DOQS_MINIMAL_BUILD=ON` flag to the `liboqs` CMake command.
This is the definitive fix for the "Argument list too long" error encountered when building on Windows/MSYS2. By enabling a minimal build, we drastically reduce the number of object files being passed to the linker, thus avoiding the command-line length limitation of the operating system.

This change is based on the official `liboqs` build documentation.