This commit fixes a dependency lookup issue in the `build_pqc_server.sh` script.
The `oqs-provider` build step was failing because it could not find the `liboqs` and `OpenSSL` libraries that were previously built and installed into a local directory.

This change adds the `-DCMAKE_PREFIX_PATH` flag to the `oqs-provider` CMake command, pointing it to the local `install` directory. This allows CMake's `find_package` to correctly locate the required dependencies.